@@ -1,15 +1,31 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Incidente;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Incidente::class, function (Faker $faker) {
-    return [
-        'nombre' => $faker->unique(true)->colorName . ' ' . rand(1,60),
-        'descripcion' => $faker->boolean ? $faker->sentence : null,
-        'created_by' => rand(1,10),
-        'updated_by' => rand(1,10),
-    ];
-});
+class IncidenteFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Incidente::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'nombre' => $this->faker->unique(true)->colorName . ' ' . $this->faker->numberBetween(1,60),
+            'descripcion' => $this->faker->boolean ? $this->faker->sentence : null,
+            'created_by' => $this->faker->numberBetween(1,10),
+            'updated_by' => $this->faker->numberBetween(1,10),
+        ];
+    }
+}
