@@ -1,15 +1,31 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Vehiculo;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Vehiculo::class, function (Faker $faker) {
-    return [
-        'nombre' => 'Vehiculo #' . $faker->unique(true)->randomNumber . chr(rand(65,90)),
-        'descripcion' => $faker->text(),
-        'created_by' => rand(1,10),
-        'updated_by' => rand(1,10),
-    ];
-});
+class VehiculoFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Vehiculo::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'nombre' => 'Vehiculo #' . $this->faker->unique(true)->randomNumber . $this->faker->lexify('?'),
+            'descripcion' => $this->faker->text(),
+            'created_by' => $this->faker->numberBetween(1,10),
+            'updated_by' => $this->faker->numberBetween(1,10),
+        ];
+    }
+}
