@@ -1,8 +1,11 @@
 <?php
 
-use Illuminate\Database\Seeder;
+namespace Database\Seeders;
 
-class GuiasImpresionTableSeeder extends Seeder
+use Illuminate\Database\Seeder;
+use App\GuiaImpresion;
+
+class GuiaImpresionSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -11,9 +14,7 @@ class GuiasImpresionTableSeeder extends Seeder
      */
     public function run()
     {
-        // return factory(App\GuiaImpresion::class, 1)->create();
-
-        return App\GuiaImpresion::create([
+        return GuiaImpresion::create([
             'nombre' => 'Información completa',
             'formato_json' => json_encode([]),
             'margenes_json' => json_encode([]),
